+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 * [WAZ.Storage][New] CloudQueue getting messages operation accept message count parameter.
 * [WAZ.Storage][New] CloudTable expose and accept table continuation token.
 * [EventStore][Improvment] Consumer can start reading from the end of streams.
+* [EventStore][Improvment] Retry notification delivery timeout.
 * [EventStore][Breaking] Notification listener starts reading stream from the end.
 * [EventStore][BugFix] Consumer initialize reader only in leader state.